requirements.txt
streamlit
stellar-sdk
pandas
matplotlib
plotly
pytest
requests
pyyaml
tzlocal
